Conquerors2-main 2/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.6
PyJWT==2.8.0
Werkzeug==2.3.7
python-socketio==5.10.0
python-engineio==4.8.0


# AI SDKs
google-generativeai==0.5.0
python-dotenv==1.0.0
